General summary
1) NPN install post-script has an error
2) More css knowledge could have been demonstrated
3) Use of Bootstrap or other framework would be good
4) Running unit tests and e2e tests don't work for me
5) Readability and maintenance concerns in e2e/unit tests
6) Lots of fragile criteria and locators used in tests, particularly in string based .html() tests
7) Various indications of cargo cult, concerns over whether the implementation detail is understood or whether code has been combined from other sources
8) Missing knowledge in HTML5/Angular APIs which has lead to a lot of duplication and complexity (and bugs)
9) No Angular templating demonstrated
10) The choice of implementation in the table directive is very peculiar.
11) Basic separation of concerns shown, which is nice!

Otherwise a complete delivery that behaves as briefed, which is very commendable!

The ideal test would pull split concerns appropriately according to Angular best practice, show experience of the correct APIs and lean on Angular's strengths.  The only bespoke javascript would be managing a collection in memory by adding/removing/sorting.  Testing would be concise and highly readable with a view to future maintenance.